+++ tests/mocks/testSchema.ts
@@ -1,6 +1,7 @@
 import { v4 as uuidv4 } from 'uuid';
 
 import type { BormSchema, DataField } from '../../src/index';
+import { isArray } from 'radash';
 //* when updating, please run `pnpm test:buildSchema`
 
 const name: DataField = {
@@ -575,6 +576,18 @@ export const testSchema: BormSchema = {
 								message: 'Name must not exist, or be less than 15 characters',
 							},
 							{
+								type: 'validate',
+								fn: ({ fields }) => {
+									if (!fields) {
+										return true;
+									}
+									fields.some((f: any) => f.name === 'forbiddenName');
+									throw new Error("You can't have a field named 'forbiddenName'");
+								}, //in general this would be run at the attribute level instead, as we use a single attribute, but is for testing
+								severity: 'error',
+								message: 'Name must not exist, or be less than 15 characters',
+							},
+							{
 								type: 'transform',
 								fn: ({ name }) =>
 									name === 'secretName'
@@ -674,6 +687,26 @@ export const testSchema: BormSchema = {
 			dataFields: [{ ...id }],
 			roles: {
 				hookTypeA: { cardinality: 'ONE' },
+			},
+			hooks: {
+				pre: [
+					{
+						triggers: {
+							onCreate: () => true,
+							onUpdate: () => true,
+						},
+						actions: [
+							{
+								type: 'validate',
+								fn: ({ hookTypeA }) => {
+									return !isArray(hookTypeA);
+								},
+								severity: 'error',
+								message: "Can't be an array",
+							},
+						],
+					},
+				],
 			},
 		},
 	},