+++ changelog.md
@@ -13,6 +13,11 @@
 - Chore: Immer downgraded to 9.0.21 (until <https://github.com/immerjs/immer/issues/1087> is fixed)
 - Fix: lot of fixes
 
+## 0.8.11(2023-02-15)
+
+- Tests: Added nested object validation test
+
+
 ## 0.8.10(2023-02-14)
 
 - Fix: Sanitized params in fn validations and transforms